Moved exception under exception folder.
